Revert "Merge branch 'ble-web' of https://github.com/MarcoMattiuz/Embedded-wearable-project into ble-web"
This reverts commit 96234b133c664a17a83326e38beee237b4a5a3f9, reversing
changes made to 640fc36efd714c396da6fe19113d109c7d3f2364.

+++ lib/MAX30102/hr_logic.h
@@ -13,6 +13,6 @@ int16_t     get_RED_AC(uint32_t sample);
 int16_t     get_IR_AC(uint32_t sample);
 int16_t     get_IR_AC2(int32_t x);
 bool        beat_detected(int16_t ir_ac);
-void        calculateBPM(int16_t ir_ac_curr,int16_t *BPM,int16_t *AVG_BPM);
+void        calculateBPM(int16_t ir_ac_curr,float *BPM,float *AVG_BPM);
 
 #endif /* HR_LOGIC_H_ */
